#26089 Added better interaction when an application is not found.
This implements better feedback when an application is not launched
corretly. For the shotgun engine, it displays a native error popup with
embedded html links to documentation. For engines with a UI, it pops up
a dialog with a button to jump to documentation. For the non-ui case,
the copy has been improved. This will improve general onramp and the
desktop workflows in particular. 

Closes #11