#33143 Better error message when flame is not found.
Previously, if the path to flame wasn't found, an error message was
written to the desktop logs which was hard to find. With this change, a
message box pops up similar to the one being shown for other apps. The
reason things are a little bit different for Flame is because the launch
sequence is more complex.

Closes #24