For #35869: do version token replacement in the args setting
Apply our version token replacement to the *_args setting like we currently do for the path and icon settings. This enables clients who use a wrapper to launch their DCC, take advantage of the version tokens in the *_args setting. And allows them to support launching multiple versions of their DCC using the `versions` setting.