For #38944, Introduce Software Entity support and general refactor
Refactored to launch DCC applications from both Project configuration and Software site entities. No changes were made to the Project configuration option, so that works as it always has. As part of the Zero Config effort, Software entities can be created on a site and populated with the information required to launch the represented DCC application. For this version of the app, a Software entity is a configurable CustomNonProjectEntity with the following custom fields:

   * sg_engine (text)
   * sg_group_restrictions (Group multi-entity)
   * sg_linux_args (text)
   * sg_linux_path (text)
   * sg_mac_args (text)
   * sg_mac_path (text)
   * sg_projects (Project multi-entity)
   * sg_user_restrictions (Person multi-entity)
   * sg_versions (text)
   * sg_windows_args (text)
   * sg_windows_path (text)

These field names will change once the Software entity becomes native in Shotgun.

Closes #34.