#40549 Implement registered command grouping (#42)
Adds `group` and `group_default` configuration settings to specify a group a launch command should be a member of and whether this launch command should run when the group is selected. This information is also stored as fields on Software entities, so the relevant information can be parsed from there. Desktop implements this grouping mechanism the same way collapse rules work. Each command in the group will appear as a drop-down menu item from the "Group" button presented on the Project page. Ideally, whichever command is specified as group_default will run when the Group button is pressed, however that requires additional updates to tk-desktop. For now, the highest version of the commands in the group will run instead.