#40550 Tweak to default app launch hook to support bundles and executables
Inserts a `-a` flag into the OSX open command to treat the supplied app path
as an application. This makes the app launching work with both application
bundles and full executable paths. Prior to this, if you included the full
executable path, a new terminal would be spawned to launch the DCC.