For #44069: Adds a new hook: before_register_command.py (#53)
The default implementation is for the hook method to be a no-op. It is intended to be overridden at the config level when redirecting to an engine instance other than what is set in the associated Software entity. A realistic example would be for Nuke and Nuke Studio. They share a single Software entity in SG, but in classic SGTK project setups Nuke Studio is configured differently than Nuke. In that situation, we need to redirect to a tk-nukestudio engine instance prior to the launcher command being registered with the engine.