For #44763: Detects Flame launches in the case of a Software entity launch. (#57)
In the event we're launching Flame classic integration via a Software entity launcher, we don't have access to the engine name in the app launch hook. As such, we now provide a new hack on top of the old one that detects that we're launching Flame classic by looking at the launch command to be executed.